API/debug_imports.py
print("Testing imports...")

try:
    from models import QueryRequest, QueryResponse
    print("OK models imported successfully")
except Exception as e:
    print(f"ERROR models import failed: {e}")

try:
    from search import find_best_answer
    print("OK search imported successfully")
except Exception as e:
    print(f"ERROR search import failed: {e}")

try:
    from llm_refiner import refine_with_gemini
    print("OK llm_refiner imported successfully")
except Exception as e:
    print(f"ERROR llm_refiner import failed: {e}")

print("Import testing completed!")